ExperimentalDesign.jl Documentation

ExperimentalDesign.jl Documentation

ExperimentalDesign.d_efficiency_lower_bound — Method.

d_efficiency_lower_bound(model_matrix::Array{Float64, 2})

Compute a lower bound for the $D$-efficiency of a given design's model matrix according to Castillo's "Process Optimization : A Statistical Approach".

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = plackett_burman(4)
4×3 Array{Int64,2}:
  1   1   1
  1  -1  -1
 -1  -1   1
 -1   1  -1

julia> factors = OrderedDict([(:f1, [-1., 1.]), (:f2, [-1., 1.]), (:f3, [-1., 1.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-1.0, 1.0]
  :f2 => [-1.0, 1.0]
  :f3 => [-1.0, 1.0]

julia> m = generate_model_matrix(@formula(y ~ f1 + f2 + f3), float(A), factors)
4×4 Array{Float64,2}:
 1.0   1.0   1.0   1.0
 1.0   1.0  -1.0  -1.0
 1.0  -1.0  -1.0   1.0
 1.0  -1.0   1.0  -1.0

julia> d_efficiency_lower_bound(m)
1.0

Formula

For a design $A_{n,p}$ with $n$ experiments or rows, $p$ factors or columns, and model matrix $\mathbf{X}$, the lower bound for the $D$-efficiency of $A$ is:

\[D_{eff}^{(L)} = \dfrac{|\mathbf{X}^{\prime}\mathbf{X}|^{1/p}}{n}\]

source

ExperimentalDesign.d_optimality — Method.

d_optimality(model_matrix::Array{Float64, 2})

Compute the $D$-optimality of a given design's model matrix.

Examples

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = plackett_burman(4)
4×3 Array{Int64,2}:
  1   1   1
  1  -1  -1
 -1  -1   1
 -1   1  -1

julia> factors = OrderedDict([(:f1, [-1., 1.]), (:f2, [-1., 1.]), (:f3, [-1., 1.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-1.0, 1.0]
  :f2 => [-1.0, 1.0]
  :f3 => [-1.0, 1.0]

julia> m = generate_model_matrix(@formula(y ~ f1 + f2 + f3), float(A), factors)
4×4 Array{Float64,2}:
 1.0   1.0   1.0   1.0
 1.0   1.0  -1.0  -1.0
 1.0  -1.0  -1.0   1.0
 1.0  -1.0   1.0  -1.0

julia> d_optimality(m)
256.0

Formula

The $D$-optimality of a design is the determinant of the information matrix $\mathbf{X}^{\prime}\mathbf{X}$, where $\mathbf{X}$ is the model matrix of a design.

source

ExperimentalDesign.expand_factors — Method.

expand_factors(factors::OrderedDict)

Return an OrderedDict with true factors expanded to [0., 1.] 2-level numerical factors.

For a true factor with $n$ levels, creates $n - 1$ 2-level factors. Only one of these new factors can be at level 1. at each row of a design, encoding each level of the original factor. The first level of the original factor is encoded by all new factors being at level 0..

Examples

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = OrderedDict(:A => [1., 2.], :B => [1, 2, 3, 4], :C => ["A", "B"], :D => [1.2, 2.2, 3.4], :E => [:a, :b, :c, :d])
DataStructures.OrderedDict{Symbol,Any} with 5 entries:
  :A => [1.0, 2.0]
  :B => [1, 2, 3, 4]
  :C => String["A", "B"]
  :D => [1.2, 2.2, 3.4]
  :E => Symbol[:a, :b, :c, :d]

julia> expand_factors(A)
DataStructures.OrderedDict{Symbol,Any} with 7 entries:
  :A   => [1.0, 2.0]
  :B   => [1, 2, 3, 4]
  :C_1 => [0.0, 1.0]
  :D   => [1.2, 2.2, 3.4]
  :E_1 => [0.0, 1.0]
  :E_2 => [0.0, 1.0]
  :E_3 => [0.0, 1.0]

source

ExperimentalDesign.generate_model_matrix — Method.

generate_model_matrix(formula::Formula,
                      design::Array{Float64, 2},
                      factors::OrderedDict;
                      scale::Function = scale_boxdraper_encoding!)

Generate a DataFrame with a scaled model matrix for a given formula, design and factors.

Assumes that formula is a linear relationship between all the factors in factors.

Examples

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = plackett_burman(4)
4×3 Array{Int64,2}:
  1   1   1
  1  -1  -1
 -1  -1   1
 -1   1  -1

julia> factors = OrderedDict([(:f1, [-1., 1.]), (:f2, [-1., 1.]), (:f3, [-1., 1.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-1.0, 1.0]
  :f2 => [-1.0, 1.0]
  :f3 => [-1.0, 1.0]

julia> m = generate_model_matrix(@formula(y ~ f1 + f2 + f3), float(A), factors)
4×4 Array{Float64,2}:
 1.0   1.0   1.0   1.0
 1.0   1.0  -1.0  -1.0
 1.0  -1.0  -1.0   1.0
 1.0  -1.0   1.0  -1.0

source

ExperimentalDesign.get_expanded_values — Method.

get_expanded_values(factors::OrderedDict)

Return a mapping from an original level of a factor to a column in an expanded design. Does not expand factors.

Examples

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = OrderedDict(:A => [1., 2.], :B => [1, 2, 3, 4], :C => ["A", "B"], :D => [1.2, 2.2, 3.4], :E => [:a, :b, :c, :d])
DataStructures.OrderedDict{Symbol,Any} with 5 entries:
  :A => [1.0, 2.0]
  :B => [1, 2, 3, 4]
  :C => String["A", "B"]
  :D => [1.2, 2.2, 3.4]
  :E => Symbol[:a, :b, :c, :d]

julia> get_expanded_values(A)
DataStructures.OrderedDict{Any,Any} with 2 entries:
  :C => DataStructures.OrderedDict{Any,Symbol}("B"=>:C_1)
  :E => DataStructures.OrderedDict{Any,Symbol}(:b=>:E_1,:c=>:E_2,:d=>:E_3)

source

ExperimentalDesign.scale_boxdraper_encoding! — Method.

scale_boxdraper_encoding!(design::Array{Float64, 2},
                          factors::Array{T, 1};
                          scale_denominator = true) where T <: Any

Scale factors of a design using the Box and Draper's coding convention from "Response Surfaces, Mixtures, and Ridge Analyses".

Examples

With denominator scaling:

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = float(plackett_burman(4))
4×3 Array{Float64,2}:
  1.0   1.0   1.0
  1.0  -1.0  -1.0
 -1.0  -1.0   1.0
 -1.0   1.0  -1.0

julia> factors = OrderedDict([(:f1, [-1., 1.]), (:f2, [-1., 1.]), (:f3, [-1., 1.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-1.0, 1.0]
  :f2 => [-1.0, 1.0]
  :f3 => [-1.0, 1.0]

julia> scale_boxdraper_encoding!(A, collect(values(factors)))
4×3 Array{Float64,2}:
  1.0   1.0   1.0
  1.0  -1.0  -1.0
 -1.0  -1.0   1.0
 -1.0   1.0  -1.0

julia> all([isapprox(1.0, sqrt(sum(A[:, i] .^ 2.0) / size(A, 1))) for i = 1:size(A, 2)])
true

Without denominator scaling:

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = float(plackett_burman(4))
4×3 Array{Float64,2}:
  1.0   1.0   1.0
  1.0  -1.0  -1.0
 -1.0  -1.0   1.0
 -1.0   1.0  -1.0

julia> factors = OrderedDict([(:f1, [-1., 1.]), (:f2, [-1., 1.]), (:f3, [-1., 1.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-1.0, 1.0]
  :f2 => [-1.0, 1.0]
  :f3 => [-1.0, 1.0]

julia> scale_boxdraper_encoding!(A, collect(values(factors)), scale_denominator = false)
4×3 Array{Float64,2}:
  0.5   0.5   0.5
  0.5  -0.5  -0.5
 -0.5  -0.5   0.5
 -0.5   0.5  -0.5

julia> all([isapprox(1.0, sqrt(sum(A[:, i] .^ 2.0))) for i = 1:size(A, 2)])
true

Formula

For a design $D_{n,p}$ with $n$ experiments or rows and $p$ factors or columns, scale each factor $\mathbf{x}_i$ to $\mathbf{x}_{i}^{s}$ according to:

\[\mathbf{x}_{i}^{s} = \dfrac{\mathbf{x}_i - \bar{\mathbf{x}}_i}{S_i}\]

Where $\bar{\mathbf{x}}_i$ is mean of factor definition values in the factors parameter and:

\[S_{i}^{2} = \dfrac{1}{n} \sum\limits_{j = 1}^{n}{(x_{ij} - \bar{\mathbf{x}}_i)^{2}}\]

If we pass scale_denominator = false, $S_i$ becomes:

\[S_{i}^{2} = \sum\limits_{j = 1}^{n}{(x_{ij} - \bar{\mathbf{x}}_i)^{2}}\]

source

ExperimentalDesign.scale_orthogonal! — Method.

scale_orthogonal!(design::Array{Float64, 2}, factors::Array{T, 1}) where T <: Any

Orthogonally scale and center factors of a design using design and factor limits.

Examples

julia> using ExperimentalDesign, DataStructures, StatsModels

julia> A = [5. 2. -1.; 5. 3. 0.; -5. 1. -2.]
3×3 Array{Float64,2}:
  5.0  2.0  -1.0
  5.0  3.0   0.0
 -5.0  1.0  -2.0

julia> factors = OrderedDict([(:f1, [-5., 0., 5.]), (:f2, [1., 2., 3.]), (:f3, [-2., -1., 0.])])
DataStructures.OrderedDict{Symbol,Array{Float64,1}} with 3 entries:
  :f1 => [-5.0, 0.0, 5.0]
  :f2 => [1.0, 2.0, 3.0]
  :f3 => [-2.0, -1.0, 0.0]

julia> scale_orthogonal!(A, collect(values(factors)))
3×3 Array{Float64,2}:
  1.0   0.0   0.0
  1.0   1.0   1.0
 -1.0  -1.0  -1.0

julia> A
3×3 Array{Float64,2}:
  1.0   0.0   0.0
  1.0   1.0   1.0
 -1.0  -1.0  -1.0

Formula

For a design $D_{n,p}$ with $n$ experiments or rows and $p$ factors or columns, scale each factor $\mathbf{x}_i$ to $\mathbf{x}_{i}^{s}$ according to:

\[\mathbf{x}_{i}^{s} = \dfrac{\mathbf{x}_i - \bar{M}}{\bar{M}_{def}}\]

Where $\mathbf{x}_{def}$ is the factor defined in the factors parameter and:

\[\bar{M} = (max(\mathbf{x}_i) + min(\mathbf{x}_i)) / 2\]

\[\bar{M}_{def} = (max(\mathbf{x}_{def}) - min(\mathbf{x}_{def})) / 2\]

source