API Documentation

Documentation for ExperimentalDesign.jl's API.

Contents

API Documentation

Contents

Index

API

Index

ExperimentalDesign.AbstractRandomDesign

ExperimentalDesign.AbstractResponseSurfaceDesign

ExperimentalDesign.IterableFullFactorial

ExperimentalDesign.RandomDesign

ExperimentalDesign.a_criterion

ExperimentalDesign.e_criterion

ExperimentalDesign.g_criterion

ExperimentalDesign.rd_criterion

ExperimentalDesign.t_criterion

API

ExperimentalDesign.AbstractRandomDesign — Type

abstract type AbstractRandomDesign <: AbstractDesign

ExperimentalDesign.AbstractResponseSurfaceDesign — Type

abstract type AbstractResponseSurfaceDesign <: AbstractDesign

ExperimentalDesign.IterableFullFactorial — Type

struct IterableFullFactorial <: AbstractFactorialDesign

matrix::DataFrame

factors::NamedTuple

formula::FormulaTerm

ExperimentalDesign.RandomDesign — Type

struct RandomDesign <: ExperimentalDesign.AbstractRandomDesign

matrix::DataFrame

factors::NamedTuple

formula::FormulaTerm

ExperimentalDesign.a_criterion — Method

a_criterion(model_matrix; tolerance) -> Any

Criterion of A-optimality which seeks minimum of $trace((X^T · X)^{-1})$. This criterion results in minimizing the average variance of the estimates of the regression coefficients.

Criterion metric is $\frac{p}{trace((X^T · X)^{-1}) · N}$.

ExperimentalDesign.e_criterion — Method

e_criterion(model_matrix; tolerance) -> Any

Criterion of E-optimality maximizes the minimum eigenvalue of the information matrix ($X^T · X$).

Minimization metric is $\frac{\min λ(X^T · X)}{N}$

ExperimentalDesign.g_criterion — Method

g_criterion(model_matrix; tolerance) -> Any

Criterion of G-optimality which seeks minimize the maximum entry in the diagonal of $X·(X^T · X)^{-1}·X^T$. This criterion results in minimizing the maximum variance of the predicted values.

Minimization metric is $\frac{N}{\max diag(H)}$ where $H = X·(X^T · X)^{-1}·X^T$.

ExperimentalDesign.rd_criterion — Method

rd_criterion(model_matrix; tolerance) -> Any

Criterion of rotation distance optimality. Experimental.

ExperimentalDesign.t_criterion — Method

t_criterion(model_matrix; tolerance) -> Any

Criterion of T-optimality. This criterion maximize $trace(X^T · X)$.

Minimization metric is $\frac{trace(X^T · X)}{N · p}$.