API Documentation

Documentation for ExperimentalDesign.jl's API.

Contents

API Documentation

Contents

Index

API

Index

ExperimentalDesign.AbstractRandomDesign

ExperimentalDesign.expand_grid

ExperimentalDesign.linear_model

ExperimentalDesign.measure_optimize_design

ExperimentalDesign.measure_optimize_design_refresh

ExperimentalDesign.variance_prediction!

API

ExperimentalDesign.AbstractRandomDesign — Type

abstract type AbstractRandomDesign <: AbstractDesign

ExperimentalDesign.expand_grid — Method

expand_grid(factors::Int64, levels::Int64) -> Array{Float64,2}

Generates all combinations of factors factors with levels levels.

ExperimentalDesign.linear_model — Method

linear_model(data::DataFrames.DataFrame) -> StatsModels.FormulaTerm{StatsModels.Term,_A} where _A

ExperimentalDesign.measure_optimize_design — Method

measure_optimize_design(levels::Any) -> Any

ExperimentalDesign.measure_optimize_design_refresh — Method

measure_optimize_design_refresh(refresh_candidate_set::Any) -> Any

ExperimentalDesign.variance_prediction! — Method

variance_prediction!(dispersion_matrix::Array{Float64,2}, model_matrix::Array{Float64,2}, candidate_set::Array{Float64,2}, design_variance::Array{Float64,1}, candidate_variance::Array{Float64,1}, predictions::Array{Float64,2}, deltas::Array{Float64,2})

Computes deltas for the exchanges between all pairs made between experiments in model_matrix and experiments in candidate_set. The dispersion_matrix array must contain the dispersion matrix for model_matrix. Both model_matrix and candidate_set must be model matrices. The arrays design_variance, candidate_variance, predictions and deltas will be overwritten.