Public

API Documentation

Documentation for ExperimentalDesign.jl's API.

Contents

API Documentation

Contents

Index

API

Index

ExperimentalDesign.AbstractDesign

ExperimentalDesign.AbstractFactorialDesign

ExperimentalDesign.AbstractOptimalDesign

ExperimentalDesign.AbstractScreeningDesign

ExperimentalDesign.FractionalFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.Optimal

ExperimentalDesign.PlackettBurman

ExperimentalDesign.PlackettBurman

ExperimentalDesign.PlackettBurman

ExperimentalDesign.d_criterion

ExperimentalDesign.expanded_design

ExperimentalDesign.explicit_fullfactorial

ExperimentalDesign.explicit_fullfactorial

ExperimentalDesign.fullfactorial

ExperimentalDesign.isplackettburman

ExperimentalDesign.next_offset_divisible_prime

ExperimentalDesign.optimize_design

ExperimentalDesign.paley

ExperimentalDesign.plackettburman

ExperimentalDesign.random_design

API

ExperimentalDesign.AbstractDesign — Type.

abstract type AbstractDesign

source

ExperimentalDesign.AbstractFactorialDesign — Type.

abstract type AbstractFactorialDesign <: AbstractDesign

source

ExperimentalDesign.AbstractOptimalDesign — Type.

abstract type AbstractOptimalDesign <: AbstractDesign

source

ExperimentalDesign.AbstractScreeningDesign — Type.

abstract type AbstractScreeningDesign <: AbstractDesign

source

ExperimentalDesign.FractionalFactorial — Type.

struct FractionalFactorial <: AbstractFactorialDesign

source

ExperimentalDesign.FullFactorial — Type.

struct FullFactorial <: AbstractFactorialDesign

Encapsulates a full factorial design. Can contain an explicit design where all experiments are completely computed and stored, or an iterator of a design, from which experiments may be extracted as needed.

matrix::Union{Missing, DataFrame}

iterator::Base.Iterators.ProductIterator

factors::NamedTuple

formula::StatsModels.FormulaTerm

source

ExperimentalDesign.FullFactorial — Method.

FullFactorial(factors::NamedTuple, formula::StatsModels.FormulaTerm; explicit) -> FullFactorial

julia> FullFactorial((A = [1, 2, 4], B = [:a, :b], C = [1.0, -1.0]), @formula(y ~ A + B +C), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ A   │ B   │ C    │
│     │ Any │ Any │ Any  │
├─────┼─────┼─────┼──────┤
│ 1   │ 1   │ a   │ 1.0  │
│ 2   │ 2   │ a   │ 1.0  │
│ 3   │ 4   │ a   │ 1.0  │
│ 4   │ 1   │ b   │ 1.0  │
│ 5   │ 2   │ b   │ 1.0  │
│ 6   │ 4   │ b   │ 1.0  │
│ 7   │ 1   │ a   │ -1.0 │
│ 8   │ 2   │ a   │ -1.0 │
│ 9   │ 4   │ a   │ -1.0 │
│ 10  │ 1   │ b   │ -1.0 │
│ 11  │ 2   │ b   │ -1.0 │
│ 12  │ 4   │ b   │ -1.0 │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (A = [1, 2, 4], B = Symbol[:a, :b], C = [1.0, -1.0]), y ~ A + B + C)

source

ExperimentalDesign.FullFactorial — Method.

FullFactorial(factors::NamedTuple; explicit) -> FullFactorial

julia> FullFactorial((A = [1, 2, 4], B = [:a, :b], C = [1.0, -1.0]), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ A   │ B   │ C    │
│     │ Any │ Any │ Any  │
├─────┼─────┼─────┼──────┤
│ 1   │ 1   │ a   │ 1.0  │
│ 2   │ 2   │ a   │ 1.0  │
│ 3   │ 4   │ a   │ 1.0  │
│ 4   │ 1   │ b   │ 1.0  │
│ 5   │ 2   │ b   │ 1.0  │
│ 6   │ 4   │ b   │ 1.0  │
│ 7   │ 1   │ a   │ -1.0 │
│ 8   │ 2   │ a   │ -1.0 │
│ 9   │ 4   │ a   │ -1.0 │
│ 10  │ 1   │ b   │ -1.0 │
│ 11  │ 2   │ b   │ -1.0 │
│ 12  │ 4   │ b   │ -1.0 │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (A = [1, 2, 4], B = Symbol[:a, :b], C = [1.0, -1.0]), response ~ A + B + C)

source

ExperimentalDesign.FullFactorial — Method.

FullFactorial(factors::Tuple; explicit) -> FullFactorial

julia> FullFactorial(([1, 2, 4], [:a, :b], [1.0, -1.0]), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ factor1 │ factor2 │ factor3 │
│     │ Any     │ Any     │ Any     │
├─────┼─────────┼─────────┼─────────┤
│ 1   │ 1       │ a       │ 1.0     │
│ 2   │ 2       │ a       │ 1.0     │
│ 3   │ 4       │ a       │ 1.0     │
│ 4   │ 1       │ b       │ 1.0     │
│ 5   │ 2       │ b       │ 1.0     │
│ 6   │ 4       │ b       │ 1.0     │
│ 7   │ 1       │ a       │ -1.0    │
│ 8   │ 2       │ a       │ -1.0    │
│ 9   │ 4       │ a       │ -1.0    │
│ 10  │ 1       │ b       │ -1.0    │
│ 11  │ 2       │ b       │ -1.0    │
│ 12  │ 4       │ b       │ -1.0    │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (factor1 = [1, 2, 4], factor2 = Symbol[:a, :b], factor3 = [1.0, -1.0]), response ~ factor1 + factor2 + factor3)

source

ExperimentalDesign.Optimal — Type.

struct Optimal <: AbstractOptimalDesign

source

ExperimentalDesign.PlackettBurman — Type.

struct PlackettBurman <: AbstractScreeningDesign

Encapsulates a Plackett-Burman screening design constructed using Paley's method. Factor levels are encoded as :high and :low symbols, and extra dummy variables, possibly aliasing interactions between factors, will be added to pad a design.

matrix::DataFrames.DataFrame

factors::Tuple

dummy_factors::Tuple

formula::StatsModels.FormulaTerm

source

ExperimentalDesign.PlackettBurman — Method.

PlackettBurman(factors::Int64) -> PlackettBurman

julia> PlackettBurman(4)
PlackettBurman(8×7 DataFrames.DataFrame
│ Row │ factor1 │ factor2 │ factor3 │ factor4 │ dummy1 │ dummy2 │ dummy3 │
│     │ Symbol  │ Symbol  │ Symbol  │ Symbol  │ Symbol │ Symbol │ Symbol │
├─────┼─────────┼─────────┼─────────┼─────────┼────────┼────────┼────────┤
│ 1   │ high    │ high    │ high    │ high    │ high   │ high   │ high   │
│ 2   │ low     │ high    │ low     │ high    │ high   │ low    │ low    │
│ 3   │ high    │ low     │ high    │ high    │ low    │ low    │ low    │
│ 4   │ low     │ high    │ high    │ low     │ low    │ low    │ high   │
│ 5   │ high    │ high    │ low     │ low     │ low    │ high   │ low    │
│ 6   │ high    │ low     │ low     │ low     │ high   │ low    │ high   │
│ 7   │ low     │ low     │ low     │ high    │ low    │ high   │ high   │
│ 8   │ low     │ low     │ high    │ low     │ high   │ high   │ low    │, (:factor1, :factor2, :factor3, :factor4), (:dummy1, :dummy2, :dummy3), response ~ factor1 + factor2 + factor3 + factor4 + dummy1 + dummy2 + dummy3)

source

ExperimentalDesign.PlackettBurman — Method.

PlackettBurman(formula::StatsModels.FormulaTerm) -> PlackettBurman

julia> PlackettBurman(@formula(y ~ x1 + x2 + x3 + x4))
PlackettBurman(8×7 DataFrames.DataFrame
│ Row │ x1     │ x2     │ x3     │ x4     │ dummy1 │ dummy2 │ dummy3 │
│     │ Symbol │ Symbol │ Symbol │ Symbol │ Symbol │ Symbol │ Symbol │
├─────┼────────┼────────┼────────┼────────┼────────┼────────┼────────┤
│ 1   │ high   │ high   │ high   │ high   │ high   │ high   │ high   │
│ 2   │ low    │ high   │ low    │ high   │ high   │ low    │ low    │
│ 3   │ high   │ low    │ high   │ high   │ low    │ low    │ low    │
│ 4   │ low    │ high   │ high   │ low    │ low    │ low    │ high   │
│ 5   │ high   │ high   │ low    │ low    │ low    │ high   │ low    │
│ 6   │ high   │ low    │ low    │ low    │ high   │ low    │ high   │
│ 7   │ low    │ low    │ low    │ high   │ low    │ high   │ high   │
│ 8   │ low    │ low    │ high   │ low    │ high   │ high   │ low    │, (:x1, :x2, :x3, :x4), (:dummy1, :dummy2, :dummy3), y ~ x1 + x2 + x3 + x4 + dummy1 + dummy2 + dummy3)

source

ExperimentalDesign.d_criterion — Method.

d_criterion(model_matrix::Array{Float64,2}) -> Float64

source

ExperimentalDesign.expanded_design — Method.

expanded_design(factors::Int64, levels::Int64) -> DataFrames.DataFrame

Generates a design from all combinations of factors factors and levels levels.

source

ExperimentalDesign.explicit_fullfactorial — Method.

explicit_fullfactorial(iterator::Base.Iterators.ProductIterator) -> Any

Receives a Base.Iterators.ProductIterator and computes an explicit full factorial design. The generated array is exponentially large.

julia> explicit_fullfactorial(fullfactorial(([-1, 1], [:a, :b, :c])))
6×2 Array{Any,2}:
 -1  :a
  1  :a
 -1  :b
  1  :b
 -1  :c
  1  :c

source

ExperimentalDesign.explicit_fullfactorial — Method.

explicit_fullfactorial(factors::Tuple) -> Any

Receives a tuple of arrays representing categorical factor levels, and computes an explicit full factorial design. The generated array is exponentially large.

julia> explicit_fullfactorial(([-1, 1], [:a, :b, :c], [1, 2]))
12×3 Array{Any,2}:
 -1  :a  1
  1  :a  1
 -1  :b  1
  1  :b  1
 -1  :c  1
  1  :c  1
 -1  :a  2
  1  :a  2
 -1  :b  2
  1  :b  2
 -1  :c  2
  1  :c  2

source

ExperimentalDesign.fullfactorial — Method.

fullfactorial(factors::Tuple) -> Base.Iterators.ProductIterator{_1} where _1

Receives a tuple of arrays representing categorical factor levels, and returns a Base.Iterators.ProductIterator. This allows full factorial designs to be arbitrarily large and only be computed as needed. To compute an explicit full factorial design, use explicit_fullfactorial.

julia> fullfactorial(Tuple([-1, 1] for i = 1:10))
Base.Iterators.ProductIterator{NTuple{10,Array{Int64,1}}}(([-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1]))

source

ExperimentalDesign.isplackettburman — Method.

isplackettburman(d::Array{Int64,2}) -> Bool

To check if a given design is a Plackett-Burman design, we must check for the following properties, obtained in the original Plackett-Burman paper:

Each component is replicated at each of its values the same number of times, that is, the sum of elements in each column is zero

Each pair of components occur together at every combination of values the same number of times, that is, the sum of each pair of columns will produce a column with the same number of occurrences of $2$ and $-2$, and twice that number of occurrences of $0$

Plackett, R.L. and Burman, J.P., 1946. The design of optimum multifactorial experiments. Biometrika, 33(4), pp.305-325.

julia> isplackettburman(plackettburman(2))
true

julia> isplackettburman(plackettburman(4))
true

julia> isplackettburman(plackettburman(16))
true

julia> isplackettburman(rand(4,4))
ERROR: MethodError: no method matching isplackettburman(::Array{Float64,2})
[...]

julia> isplackettburman(rand(Int, 4,4))
false

source

ExperimentalDesign.next_offset_divisible_prime — Method.

next_offset_divisible_prime(n::Int64, offset::Int64, divisor::Int64, tries::Int64) -> Int64

Gets the next prime p, starting from n, for which (p + offset) % divisor == 0 holds.

julia> next_offset_divisible_prime(3, 1, 4, 1000)
3

julia> next_offset_divisible_prime(5, 1, 4, 1000)
7

julia> next_offset_divisible_prime(4, 1, 4, 1000)
7

source

ExperimentalDesign.optimize_design — Method.

optimize_design(; factors, levels, experiments, design, candidate_set, iterations, refresh_candidate_set)

source

ExperimentalDesign.paley — Method.

paley(matrix::Array{Int64,2}) -> Array{Int64,2}

The Paley construction is a method for constructing Hadamard matrices using finite fields. See https://en.wikipedia.org/wiki/Paley_construction

julia> paley(Matrix{Int}(undef, 8, 8))
8×8 Array{Int64,2}:
 -1   1   1  -1   1   1   1  -1
  1   1  -1   1   1   1  -1  -1
  1  -1   1   1   1  -1  -1   1
 -1   1   1   1  -1  -1   1   1
  1   1   1  -1  -1   1   1  -1
  1   1  -1  -1   1   1  -1   1
  1  -1  -1   1   1  -1   1   1
 -1  -1   1   1  -1   1   1   1

source

ExperimentalDesign.plackettburman — Method.

plackettburman(matrix_size::Int64) -> Array{Int64,2}

Constructs a Plackett-Burman design with size matrix_size if possible, or to the closest, largest, number for which it is possible.

julia> plackettburman(4)
8×7 Array{Int64,2}:
  1   1   1   1   1   1   1
 -1   1  -1   1   1  -1  -1
  1  -1   1   1  -1  -1  -1
 -1   1   1  -1  -1  -1   1
  1   1  -1  -1  -1   1  -1
  1  -1  -1  -1   1  -1   1
 -1  -1  -1   1  -1   1   1
 -1  -1   1  -1   1   1  -1

source

ExperimentalDesign.random_design — Method.

random_design(factors::Int64, size::Int64) -> DataFrames.DataFrame

Generates a random design of size size with factors numerical factors.

source