API Documentation

Documentation for ExperimentalDesign.jl's API.

Contents

API Documentation

Contents

Index

API

Index

ExperimentalDesign.AbstractDesign

ExperimentalDesign.AbstractFactorialDesign

ExperimentalDesign.AbstractOptimalDesign

ExperimentalDesign.AbstractScreeningDesign

ExperimentalDesign.FractionalFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.FullFactorial

ExperimentalDesign.OptimalDesign

ExperimentalDesign.PlackettBurman

ExperimentalDesign.PlackettBurman

ExperimentalDesign.PlackettBurman

ExperimentalDesign.RandomDesign

ExperimentalDesign.RandomDesign

ExperimentalDesign.RandomDesign

ExperimentalDesign.RandomDesign

ExperimentalDesign.RandomDesign

Base.rand

ExperimentalDesign.d_criterion

ExperimentalDesign.expanded_design

ExperimentalDesign.explicit_fullfactorial

ExperimentalDesign.explicit_fullfactorial

ExperimentalDesign.fullfactorial

ExperimentalDesign.generate_random_design

ExperimentalDesign.isplackettburman

ExperimentalDesign.next_offset_divisible_prime

ExperimentalDesign.optimize_design

ExperimentalDesign.paley

ExperimentalDesign.plackettburman

ExperimentalDesign.random_design

API

ExperimentalDesign.AbstractDesign — Type

abstract type AbstractDesign

ExperimentalDesign.AbstractFactorialDesign — Type

abstract type AbstractFactorialDesign <: AbstractDesign

ExperimentalDesign.AbstractOptimalDesign — Type

abstract type AbstractOptimalDesign <: AbstractDesign

ExperimentalDesign.AbstractScreeningDesign — Type

abstract type AbstractScreeningDesign <: AbstractDesign

ExperimentalDesign.FractionalFactorial — Type

struct FractionalFactorial <: AbstractFactorialDesign

ExperimentalDesign.FullFactorial — Type

struct FullFactorial <: AbstractFactorialDesign

Encapsulates a full factorial design. Can contain an explicit design where all experiments are completely computed and stored, or an iterator of a design, from which experiments may be extracted as needed.

matrix::Union{Missing, DataFrames.DataFrame}

iterator::Base.Iterators.ProductIterator

factors::NamedTuple

formula::StatsModels.FormulaTerm

ExperimentalDesign.FullFactorial — Method

FullFactorial(factors::Array; explicit) -> FullFactorial

julia> FullFactorial(fill([-1, 1], 3), explicit = true)
FullFactorial(8×3 DataFrames.DataFrame
│ Row │ factor1 │ factor2 │ factor3 │
│     │ Int64   │ Int64   │ Int64   │
├─────┼─────────┼─────────┼─────────┤
│ 1   │ -1      │ -1      │ -1      │
│ 2   │ 1       │ -1      │ -1      │
│ 3   │ -1      │ 1       │ -1      │
│ 4   │ 1       │ 1       │ -1      │
│ 5   │ -1      │ -1      │ 1       │
│ 6   │ 1       │ -1      │ 1       │
│ 7   │ -1      │ 1       │ 1       │
│ 8   │ 1       │ 1       │ 1       │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Int64,1},Array{Int64,1}}}(([-1, 1], [-1, 1], [-1, 1])), (factor1 = [-1, 1], factor2 = [-1, 1], factor3 = [-1, 1]), response ~ factor1 + factor2 + factor3)

ExperimentalDesign.FullFactorial — Method

FullFactorial(factors::NamedTuple, formula::StatsModels.FormulaTerm; explicit) -> FullFactorial

julia> FullFactorial((A = [1, 2, 4], B = [:a, :b], C = [1.0, -1.0]), @formula(y ~ A + B +C), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ A   │ B   │ C    │
│     │ Any │ Any │ Any  │
├─────┼─────┼─────┼──────┤
│ 1   │ 1   │ a   │ 1.0  │
│ 2   │ 2   │ a   │ 1.0  │
│ 3   │ 4   │ a   │ 1.0  │
│ 4   │ 1   │ b   │ 1.0  │
│ 5   │ 2   │ b   │ 1.0  │
│ 6   │ 4   │ b   │ 1.0  │
│ 7   │ 1   │ a   │ -1.0 │
│ 8   │ 2   │ a   │ -1.0 │
│ 9   │ 4   │ a   │ -1.0 │
│ 10  │ 1   │ b   │ -1.0 │
│ 11  │ 2   │ b   │ -1.0 │
│ 12  │ 4   │ b   │ -1.0 │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (A = [1, 2, 4], B = Symbol[:a, :b], C = [1.0, -1.0]), y ~ A + B + C)

ExperimentalDesign.FullFactorial — Method

FullFactorial(factors::NamedTuple; explicit) -> FullFactorial

julia> FullFactorial((A = [1, 2, 4], B = [:a, :b], C = [1.0, -1.0]), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ A   │ B   │ C    │
│     │ Any │ Any │ Any  │
├─────┼─────┼─────┼──────┤
│ 1   │ 1   │ a   │ 1.0  │
│ 2   │ 2   │ a   │ 1.0  │
│ 3   │ 4   │ a   │ 1.0  │
│ 4   │ 1   │ b   │ 1.0  │
│ 5   │ 2   │ b   │ 1.0  │
│ 6   │ 4   │ b   │ 1.0  │
│ 7   │ 1   │ a   │ -1.0 │
│ 8   │ 2   │ a   │ -1.0 │
│ 9   │ 4   │ a   │ -1.0 │
│ 10  │ 1   │ b   │ -1.0 │
│ 11  │ 2   │ b   │ -1.0 │
│ 12  │ 4   │ b   │ -1.0 │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (A = [1, 2, 4], B = Symbol[:a, :b], C = [1.0, -1.0]), response ~ A + B + C)

ExperimentalDesign.FullFactorial — Method

FullFactorial(factors::Tuple; explicit) -> FullFactorial

julia> FullFactorial(([1, 2, 4], [:a, :b], [1.0, -1.0]), explicit = true)
FullFactorial(12×3 DataFrames.DataFrame
│ Row │ factor1 │ factor2 │ factor3 │
│     │ Any     │ Any     │ Any     │
├─────┼─────────┼─────────┼─────────┤
│ 1   │ 1       │ a       │ 1.0     │
│ 2   │ 2       │ a       │ 1.0     │
│ 3   │ 4       │ a       │ 1.0     │
│ 4   │ 1       │ b       │ 1.0     │
│ 5   │ 2       │ b       │ 1.0     │
│ 6   │ 4       │ b       │ 1.0     │
│ 7   │ 1       │ a       │ -1.0    │
│ 8   │ 2       │ a       │ -1.0    │
│ 9   │ 4       │ a       │ -1.0    │
│ 10  │ 1       │ b       │ -1.0    │
│ 11  │ 2       │ b       │ -1.0    │
│ 12  │ 4       │ b       │ -1.0    │, Base.Iterators.ProductIterator{Tuple{Array{Int64,1},Array{Symbol,1},Array{Float64,1}}}(([1, 2, 4], Symbol[:a, :b], [1.0, -1.0])), (factor1 = [1, 2, 4], factor2 = Symbol[:a, :b], factor3 = [1.0, -1.0]), response ~ factor1 + factor2 + factor3)

ExperimentalDesign.OptimalDesign — Type

struct OptimalDesign <: AbstractOptimalDesign

matrix::DataFrames.DataFrame

candidates::AbstractDesign

criteria::Dict{Symbol,Float64}

formula::StatsModels.FormulaTerm

ExperimentalDesign.PlackettBurman — Type

struct PlackettBurman <: AbstractScreeningDesign

Encapsulates a Plackett-Burman screening design constructed using Paley's method. Factor levels are encoded as :high and :low symbols, and extra dummy variables, possibly aliasing interactions between factors, will be added to pad a design.

matrix::DataFrames.DataFrame

factors::Tuple

dummy_factors::Tuple

formula::StatsModels.FormulaTerm

ExperimentalDesign.PlackettBurman — Method

PlackettBurman(factors::Int64) -> PlackettBurman

julia> PlackettBurman(4)
PlackettBurman(8×7 DataFrames.DataFrame
│ Row │ factor1 │ factor2 │ factor3 │ factor4 │ dummy1 │ dummy2 │ dummy3 │
│     │ Int64   │ Int64   │ Int64   │ Int64   │ Int64  │ Int64  │ Int64  │
├─────┼─────────┼─────────┼─────────┼─────────┼────────┼────────┼────────┤
│ 1   │ 1       │ 1       │ 1       │ 1       │ 1      │ 1      │ 1      │
│ 2   │ -1      │ 1       │ -1      │ 1       │ 1      │ -1     │ -1     │
│ 3   │ 1       │ -1      │ 1       │ 1       │ -1     │ -1     │ -1     │
│ 4   │ -1      │ 1       │ 1       │ -1      │ -1     │ -1     │ 1      │
│ 5   │ 1       │ 1       │ -1      │ -1      │ -1     │ 1      │ -1     │
│ 6   │ 1       │ -1      │ -1      │ -1      │ 1      │ -1     │ 1      │
│ 7   │ -1      │ -1      │ -1      │ 1       │ -1     │ 1      │ 1      │
│ 8   │ -1      │ -1      │ 1       │ -1      │ 1      │ 1      │ -1     │, (:factor1, :factor2, :factor3, :factor4), (:dummy1, :dummy2, :dummy3), response ~ -1 + factor1 + factor2 + factor3 + factor4 + dummy1 + dummy2 + dummy3)

ExperimentalDesign.PlackettBurman — Method

PlackettBurman(formula::StatsModels.FormulaTerm; symbol_encoding) -> PlackettBurman

julia> PlackettBurman(@formula(y ~ x1 + x2 + x3 + x4))
PlackettBurman(8×7 DataFrames.DataFrame
│ Row │ x1    │ x2    │ x3    │ x4    │ dummy1 │ dummy2 │ dummy3 │
│     │ Int64 │ Int64 │ Int64 │ Int64 │ Int64  │ Int64  │ Int64  │
├─────┼───────┼───────┼───────┼───────┼────────┼────────┼────────┤
│ 1   │ 1     │ 1     │ 1     │ 1     │ 1      │ 1      │ 1      │
│ 2   │ -1    │ 1     │ -1    │ 1     │ 1      │ -1     │ -1     │
│ 3   │ 1     │ -1    │ 1     │ 1     │ -1     │ -1     │ -1     │
│ 4   │ -1    │ 1     │ 1     │ -1    │ -1     │ -1     │ 1      │
│ 5   │ 1     │ 1     │ -1    │ -1    │ -1     │ 1      │ -1     │
│ 6   │ 1     │ -1    │ -1    │ -1    │ 1      │ -1     │ 1      │
│ 7   │ -1    │ -1    │ -1    │ 1     │ -1     │ 1      │ 1      │
│ 8   │ -1    │ -1    │ 1     │ -1    │ 1      │ 1      │ -1     │, (:x1, :x2, :x3, :x4), (:dummy1, :dummy2, :dummy3), y ~ -1 + x1 + x2 + x3 + x4 + dummy1 + dummy2 + dummy3)

ExperimentalDesign.RandomDesign — Type

struct RandomDesign <: ExperimentalDesign.AbstractRandomDesign

factors::NamedTuple

formula::StatsModels.FormulaTerm

Encapsulates a random design generator. Receives a NamedTuple of factor names associated with Distributions from the the Distributions package. After instantiating a RandomDesign, you must request samples from it using rand.

ExperimentalDesign.RandomDesign — Method

RandomDesign(factors::Array) -> RandomDesign

julia> RandomDesign([Distributions.Uniform(2, 3), Distributions.DiscreteUniform(-1, 5), Distributions.Uniform(5, 10)])
RandomDesign((factor1 = Distributions.Uniform{Float64}(a=2.0, b=3.0), factor2 = Distributions.DiscreteUniform(a=-1, b=5), factor3 = Distributions.Uniform{Float64}(a=5.0, b=10.0)), response ~ factor1 + factor2 + factor3)

ExperimentalDesign.RandomDesign — Method

RandomDesign(factors::NamedTuple) -> RandomDesign

julia> RandomDesign((f1 = Distributions.Uniform(2, 3), f2 = Distributions.DiscreteUniform(-1, 5), f3 = Distributions.Uniform(5, 10)))
RandomDesign((f1 = Distributions.Uniform{Float64}(a=2.0, b=3.0), f2 = Distributions.DiscreteUniform(a=-1, b=5), f3 = Distributions.Uniform{Float64}(a=5.0, b=10.0)), response ~ f1 + f2 + f3)

ExperimentalDesign.RandomDesign — Method

RandomDesign(factors::Tuple) -> RandomDesign

julia> RandomDesign((Distributions.Uniform(2, 3), Distributions.DiscreteUniform(-1, 5), Distributions.Uniform(5, 10)))
RandomDesign((factor1 = Distributions.Uniform{Float64}(a=2.0, b=3.0), factor2 = Distributions.DiscreteUniform(a=-1, b=5), factor3 = Distributions.Uniform{Float64}(a=5.0, b=10.0)), response ~ factor1 + factor2 + factor3)

ExperimentalDesign.RandomDesign — Method

RandomDesign(distribution::D<:Distributions.Distribution, n::D<:Distributions.Distribution)

julia> RandomDesign(DiscreteNonParametric([-1, 1], [0.5, 0.5]), 6)
RandomDesign((factor1 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5]), factor2 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5]), factor3 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5]), factor4 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5]), factor5 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5]), factor6 = Distributions.DiscreteNonParametric{Int64,Float64,Array{Int64,1},Array{Float64,1}}(support=[-1, 1], p=[0.5, 0.5])), response ~ factor1 + factor2 + factor3 + factor4 + factor5 + factor6)

Base.rand — Function

rand(design::RandomDesign) -> Any
rand(design::RandomDesign, n::Int64) -> Any

julia> rand(RandomDesign((f1 = Distributions.Uniform(2, 3), f2 = Distributions.DiscreteUniform(-1, 5), f3 = Distributions.Uniform(5, 10))), 12)
12×3 DataFrames.DataFrame
│ Row │ f1      │ f2   │ f3      │
│     │ Real    │ Real │ Real    │
├─────┼─────────┼──────┼─────────┤
│ 1   │ 2.04922 │ 5    │ 8.85741 │
│ 2   │ 2.25659 │ -1   │ 6.57617 │
│ 3   │ 2.86526 │ 4    │ 5.41422 │
│ 4   │ 2.81201 │ -1   │ 5.40944 │
│ 5   │ 2.92412 │ 1    │ 6.22902 │
│ 6   │ 2.20051 │ 5    │ 8.80749 │
│ 7   │ 2.69459 │ 3    │ 6.34626 │
│ 8   │ 2.28902 │ 2    │ 8.72759 │
│ 9   │ 2.95173 │ 0    │ 5.42517 │
│ 10  │ 2.35163 │ 4    │ 5.89413 │
│ 11  │ 2.97614 │ 5    │ 9.99911 │
│ 12  │ 2.4577  │ 0    │ 8.303   │

ExperimentalDesign.d_criterion — Method

d_criterion(model_matrix::Array{Float64,2}) -> Float64

ExperimentalDesign.expanded_design — Method

expanded_design(factors::Int64, levels::Int64) -> DataFrames.DataFrame

Generates a design from all combinations of factors factors and levels levels.

ExperimentalDesign.explicit_fullfactorial — Method

explicit_fullfactorial(iterator::Base.Iterators.ProductIterator) -> Any

Receives a Base.Iterators.ProductIterator and computes an explicit full factorial design. The generated array is exponentially large.

julia> explicit_fullfactorial(fullfactorial(([-1, 1], [:a, :b, :c])))
6×2 Array{Any,2}:
 -1  :a
  1  :a
 -1  :b
  1  :b
 -1  :c
  1  :c

ExperimentalDesign.explicit_fullfactorial — Method

explicit_fullfactorial(factors::Tuple) -> Any

Receives a tuple of arrays representing categorical factor levels, and computes an explicit full factorial design. The generated array is exponentially large.

julia> explicit_fullfactorial(([-1, 1], [:a, :b, :c], [1, 2]))
12×3 Array{Any,2}:
 -1  :a  1
  1  :a  1
 -1  :b  1
  1  :b  1
 -1  :c  1
  1  :c  1
 -1  :a  2
  1  :a  2
 -1  :b  2
  1  :b  2
 -1  :c  2
  1  :c  2

ExperimentalDesign.fullfactorial — Method

fullfactorial(factors::Tuple) -> Base.Iterators.ProductIterator

Receives a tuple of arrays representing categorical factor levels, and returns a Base.Iterators.ProductIterator. This allows full factorial designs to be arbitrarily large and only be computed as needed. To compute an explicit full factorial design, use explicit_fullfactorial.

julia> fullfactorial(Tuple([-1, 1] for i = 1:10))
Base.Iterators.ProductIterator{NTuple{10,Array{Int64,1}}}(([-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1], [-1, 1]))

ExperimentalDesign.generate_random_design — Method

generate_random_design(factors::Int64, size::Int64) -> DataFrames.DataFrame

Generates a random design of size size with factors numerical factors.

ExperimentalDesign.isplackettburman — Method

isplackettburman(d::Array{Int64,2}) -> Bool

To check if a given design is a Plackett-Burman design, we must check for the following properties, obtained in the original Plackett-Burman paper:

Each component is replicated at each of its values the same number of times, that is, the sum of elements in each column is zero

Each pair of components occur together at every combination of values the same number of times, that is, the sum of each pair of columns will produce a column with the same number of occurrences of $2$ and $-2$, and twice that number of occurrences of $0$

Plackett, R.L. and Burman, J.P., 1946. The design of optimum multifactorial experiments. Biometrika, 33(4), pp.305-325.

julia> isplackettburman(plackettburman(2))
true

julia> isplackettburman(plackettburman(4))
true

julia> isplackettburman(plackettburman(16))
true

julia> isplackettburman(rand(Int, 4,4))
false

ExperimentalDesign.next_offset_divisible_prime — Method

next_offset_divisible_prime(n::Int64, offset::Int64, divisor::Int64, tries::Int64) -> Int64

Gets the next prime p, starting from n, for which (p + offset) % divisor == 0 holds.

julia> next_offset_divisible_prime(3, 1, 4, 1000)
3

julia> next_offset_divisible_prime(5, 1, 4, 1000)
7

julia> next_offset_divisible_prime(4, 1, 4, 1000)
7

ExperimentalDesign.optimize_design — Method

optimize_design(; factors, levels, experiments, design, candidate_set, iterations, refresh_candidate_set)

ExperimentalDesign.paley — Method

paley(matrix::Array{Int64,2}) -> Array{Int64,2}

The Paley construction is a method for constructing Hadamard matrices using finite fields.

julia> paley(Matrix{Int}(undef, 8, 8))
8×8 Array{Int64,2}:
 -1   1   1  -1   1   1   1  -1
  1   1  -1   1   1   1  -1  -1
  1  -1   1   1   1  -1  -1   1
 -1   1   1   1  -1  -1   1   1
  1   1   1  -1  -1   1   1  -1
  1   1  -1  -1   1   1  -1   1
  1  -1  -1   1   1  -1   1   1
 -1  -1   1   1  -1   1   1   1

ExperimentalDesign.plackettburman — Method

plackettburman(matrix_size::Int64) -> Array{Int64,2}

Constructs a Plackett-Burman design with size matrix_size if possible, or to the closest, largest, number for which it is possible.

julia> plackettburman(4)
8×7 Array{Int64,2}:
  1   1   1   1   1   1   1
 -1   1  -1   1   1  -1  -1
  1  -1   1   1  -1  -1  -1
 -1   1   1  -1  -1  -1   1
  1   1  -1  -1  -1   1  -1
  1  -1  -1  -1   1  -1   1
 -1  -1  -1   1  -1   1   1
 -1  -1   1  -1   1   1  -1

ExperimentalDesign.random_design — Method

random_design(distributions::Tuple, size::Int64) -> Any

julia> random_design((Uniform(2, 3), DiscreteUniform(-1, 5), Uniform(5, 10)), 10)
10×3 Array{Real,2}:
 2.04922   5  8.85741
 2.25659  -1  6.57617
 2.86526   4  5.41422
 2.81201  -1  5.40944
 2.92412   1  6.22902
 2.20051   5  8.80749
 2.69459   3  6.34626
 2.28902   2  8.72759
 2.95173   0  5.42517
 2.35163   4  5.89413